+++ backend-rust/src/graphql_api/baker.rs
@@ -78,16 +78,16 @@ impl QueryBaker {
                     payday_transaction_commission as "payday_transaction_commission?",
                     payday_baking_commission as "payday_baking_commission?",
                     payday_finalization_commission as "payday_finalization_commission?",
-                    lottery_power as "lottery_power?",
-                    ranking_by_lottery_powers as "ranking_by_lottery_powers?",
-                    (SELECT MAX(ranking_by_lottery_powers) FROM bakers_lottery_powers) as "total_ranking_by_lottery_powers?",
+                    payday_lottery_power as "payday_lottery_power?",
+                    payday_ranking_by_lottery_powers as "payday_ranking_by_lottery_powers?",
+                    (SELECT MAX(payday_ranking_by_lottery_powers) FROM bakers_payday_lottery_powers) as "payday_total_ranking_by_lottery_powers?",
                     pool_total_staked,
                     pool_delegator_count
                 FROM bakers
                     LEFT JOIN bakers_payday_commission_rates
                         ON bakers_payday_commission_rates.id = bakers.id
-                    LEFT JOIN bakers_lottery_powers
-                        ON bakers_lottery_powers.id = bakers.id
+                    LEFT JOIN bakers_payday_lottery_powers
+                        ON bakers_payday_lottery_powers.id = bakers.id
                 WHERE
                     (NOT $6 OR bakers.id            > $1 AND bakers.id            < $2) AND
                     (NOT $7 OR staked               > $1 AND staked               < $2) AND
@@ -95,22 +95,6 @@ impl QueryBaker {
                     (NOT $9 OR pool_delegator_count > $1 AND pool_delegator_count < $2) AND
                     -- filters
                     ($10::pool_open_status IS NULL OR open_status = $10::pool_open_status)
-                GROUP BY 
-                    bakers.id, 
-                    staked, 
-                    restake_earnings, 
-                    open_status, 
-                    metadata_url,
-                    transaction_commission, 
-                    baking_commission, 
-                    finalization_commission,
-                    payday_transaction_commission, 
-                    payday_baking_commission, 
-                    payday_finalization_commission,
-                    lottery_power, 
-                    ranking_by_lottery_powers, 
-                    pool_total_staked, 
-                    pool_delegator_count
                 ORDER BY
                     (CASE WHEN $6 AND     $3 THEN bakers.id            END) DESC,
                     (CASE WHEN $6 AND NOT $3 THEN bakers.id            END) ASC,
@@ -254,11 +238,11 @@ pub struct Baker {
     payday_transaction_commission: Option<i64>,
     payday_baking_commission: Option<i64>,
     payday_finalization_commission: Option<i64>,
-    lottery_power: Option<BigDecimal>,
+    payday_lottery_power: Option<BigDecimal>,
+    payday_ranking_by_lottery_powers: Option<i64>,
+    payday_total_ranking_by_lottery_powers: Option<i64>,
     pool_total_staked: i64,
     pool_delegator_count: i64,
-    ranking_by_lottery_powers: Option<i64>,
-    total_ranking_by_lottery_powers: Option<i64>,
 }
 impl Baker {
     pub async fn query_by_id(pool: &PgPool, baker_id: i64) -> ApiResult<Option<Self>> {
@@ -277,31 +261,15 @@ impl Baker {
                 payday_transaction_commission as "payday_transaction_commission?",
                 payday_baking_commission as "payday_baking_commission?",
                 payday_finalization_commission as "payday_finalization_commission?",
-                lottery_power as "lottery_power?",
-                ranking_by_lottery_powers as "ranking_by_lottery_powers?",
-                (SELECT MAX(ranking_by_lottery_powers) FROM bakers_lottery_powers) as "total_ranking_by_lottery_powers?",
+                payday_lottery_power as "payday_lottery_power?",
+                payday_ranking_by_lottery_powers as "payday_ranking_by_lottery_powers?",
+                (SELECT MAX(payday_ranking_by_lottery_powers) FROM bakers_payday_lottery_powers) as "payday_total_ranking_by_lottery_powers?",
                 pool_total_staked,
                 pool_delegator_count
             FROM bakers
                 LEFT JOIN bakers_payday_commission_rates ON bakers_payday_commission_rates.id = bakers.id
-                LEFT JOIN bakers_lottery_powers ON bakers_lottery_powers.id = bakers.id
-            WHERE bakers.id = $1
-            GROUP BY 
-                bakers.id, 
-                staked, 
-                restake_earnings, 
-                open_status, 
-                metadata_url,
-                transaction_commission, 
-                baking_commission, 
-                finalization_commission,
-                payday_transaction_commission, 
-                payday_baking_commission, 
-                payday_finalization_commission,
-                lottery_power, 
-                ranking_by_lottery_powers, 
-                pool_total_staked, 
-                pool_delegator_count;
+                LEFT JOIN bakers_payday_lottery_powers ON bakers_payday_lottery_powers.id = bakers.id
+            WHERE bakers.id = $1;
             "#,
             baker_id
         )
@@ -546,8 +514,10 @@ impl Baker {
         let delegated_stake_cap = min(leverage_bound_cap_for_pool, capital_bound_cap_for_pool);
 
         // Get the rank of the baker based on its lottery power.
-
-        let rank = match (self.ranking_by_lottery_powers, self.total_ranking_by_lottery_powers) {
+        let rank = match (
+            self.payday_ranking_by_lottery_powers,
+            self.payday_total_ranking_by_lottery_powers,
+        ) {
             (Some(rank), Some(total)) => Some(Ranking {
                 rank,
                 total,
@@ -573,8 +543,8 @@ impl Baker {
                 },
                 rank,
                 payday_commission_rates,
-                lottery_power: self
-                    .lottery_power
+                payday_lottery_power: self
+                    .payday_lottery_power
                     .as_ref()
                     .unwrap_or(&BigDecimal::default())
                     .try_into()
@@ -883,14 +853,14 @@ struct BakerPool<'a> {
     payday_commission_rates: Option<CommissionRates>,
     /// The lottery power of the baker pool during the last payday period
     /// captured from the `get_election_info` node endpoint.`
-    lottery_power: Decimal,
-    /// The maximum amount that may be delegated to the pool, accounting for
-    /// leverage and capital bounds.
-    delegated_stake_cap: Amount,
+    payday_lottery_power: Decimal,
     /// Ranking of the bakers by lottery powers staring with rank 1 for the
     /// baker with the highest lottery power and ending with the rank
     /// `total` for the baker with the lowest lottery power.
     rank: Option<Ranking>,
+    /// The maximum amount that may be delegated to the pool, accounting for
+    /// leverage and capital bounds.
+    delegated_stake_cap: Amount,
     open_status: Option<BakerPoolOpenStatus>,
     metadata_url: Option<&'a str>,
     // TODO: apy(period: ApyPeriod!): PoolApy!
@@ -909,8 +879,7 @@ impl<'a> BakerPool<'a> {
     // Future improvement (API breaking changes): The front end queries
     // `ranking_by_total_stake` which was renamed and updated to
     // `ranking_by_lottery_power` in the backend. Migrate the name change to the
-    // front-end and make this field not optional anymore. They value exists because
-    // the ranking is updated every block.
+    // front-end and make this field not optional anymore.
     async fn ranking_by_total_stake(&self) -> Option<Ranking> { self.rank }
 
     async fn delegated_stake(&self) -> Amount { self.delegated_stake }
@@ -925,7 +894,7 @@ impl<'a> BakerPool<'a> {
         &self.payday_commission_rates
     }
 
-    async fn lottery_power(&self) -> &Decimal { &self.lottery_power }
+    async fn lottery_power(&self) -> &Decimal { &self.payday_lottery_power }
 
     async fn open_status(&self) -> Option<BakerPoolOpenStatus> { self.open_status }
 
@@ -1069,9 +1038,9 @@ struct DelegatedStakeBounds {
     capital_bound:              i64,
 }
 
-/// Ranking of the bakers by lottery powers staring with rank 1 for the
-/// baker with the highest lottery power and ending with the rank
-/// `total` for the baker with the lowest lottery power.
+/// Ranking of the bakers by lottery powers from the last payday block staring
+/// with rank 1 for the baker with the highest lottery power and ending with the
+/// rank `total` for the baker with the lowest lottery power.
 #[derive(SimpleObject, Clone, Copy)]
 struct Ranking {
     rank:  i64,